backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3

# Validation
pydantic==2.9.2
pydantic-settings==2.6.1

# Auth
python-jose[cryptography]==3.3.0
httpx==0.27.2

# Task queue
celery==5.4.0
redis==5.1.1

# Object storage (S3-compatible)
aioboto3==13.1.1

# PDF extraction
pdfminer.six==20231228
pypdf==4.3.1

# NLP
spacy==3.7.6
# scispaCy installed separately in Dockerfile (see comment there)

# Encryption
cryptography==43.0.3

# Notifications
sendgrid==6.11.0

# Rate limiting
slowapi==0.1.9

# Utilities
python-dateutil==2.9.0
structlog==24.4.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2
factory-boy==3.3.1
